Skeleton Code+Tests for Question Retrieval
Added some skeleton code and some rough tests for the Question Retrieval
Function.
Skeleton code includes constructors supporting default values for the
sql db and dependency injection for testing purposes.

Retrieve method which accepts an int. Currently Just redirects to a
blank page but includes the source in the viewbag for testing purposes.
This needs to be refactored as it is an unsatisfactory test [essentially
testing C# works, not the functionality of the function]

Also added a database project which can be used to create an SQL script
to generate the empty database. Then used this to create the local db
which is used in development.

Entity Framework used to generate models but these are currently unused.