Refactored into Database Proxy class
Created IQuestionInserter and IQuestionRetriever. Each responsible for
inserting and retrieving questions respectively. Then refactored the
logic from QuestionsController into a Database Proxy class, which
implements IQuestionInserter and IQuestionRetriever.

This means the logic of QuestionControllers.Insert() is now just
_questionInserter.Insert(), and similar for Retrieve().

The implementation details are left up to whatever class implements the
interfaces and are used by the developer.