added Neo4j Skeleton and Updated Database Schema
Added some bare skeleton code for the neo4j database class and updated
the database schema to include a questiontype column. Allowing for easy
determination of a questions type and expected answer.

It may be better to split this question type into a seperate table with
Question ID so that a single question can be answered as one of various
types,