Added some Frontend files and Custom WebConfig k/v
Added bootstrap and angular files through nuget.

Added a custom key/value pair in the appsettings section of the
Web.Config file. This stores the application name, currently "Learn
Yesen Smarter", which can be accessed from views/anywhere in the
codebase. This means that if the name is altered in the future it only
needs to be altered in one place rather than tracked down through out
the codebase and changed multiple times.