Added authoring plumbing code
Added some code to connect the general workflow of authoring a new task.
This means there are now views for selecting author a new task, then
selecting one of the task types and being taken to a view which
displays, currently, just the name of the type and can be adapted in the
future to show an interface specific to that type.