Refactored passing questiontype use objs/json
Refactored the code so that Objects can be used to manage questiontypes
(adding a new question type means just creating a new class and
inheriting from questiontype, meaning it's easier to extend). For this,
encoding and decoding the object into a json string was necessary, so it
could be passed from the view back to the controller.