[BUGFIX] jQuery works on the manual performance page now
Previously, the jQuery code wasn't executing when a question was
clicked.

This was due to the jQuery scripts not being rendered in the shared
layout.
This meant the jQuery code was undefined nothingness to the browser.

Now, the jQuery scripts get rendered in the shared layout used by the
performance pages so they should all have access to them. [This might
need refactoring if it turns out only one or two pages use them but I
doubt that]